Update on "Refactor stable shim linters to dedup some code"
[ghstack-poisoned]
